requirements.txt
amqp==2.4.2
billiard==3.6.0.0
celery==4.3.0
dj-database-url==0.5.0
gensim==3.6.0
google-api-core==1.8.1
google-auth==1.6.3
google-cloud==0.34.0
google-cloud-language==1.1.1
google-cloud-speech==1.0.0
google-cloud-vision==0.36.0
googleapis-common-protos==1.5.8
imageio==2.5.0
imageio-ffmpeg==0.3.0
imagesize==1.1.0
lexrankr==0.1.6
pdf2image==1.5.1
pdfminer.six==20181108
Pillow==5.4.1
pydub==0.23.1
pytube==9.5.0
urllib3==1.24.2
watson-developer-cloud==2.10.1
